2008-05-27 17:11 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/source/rtl/idle.c     * updated included header files     * removed unnecessary separated code for MacOSX     * small modification in comment
  * harbour/source/hbpcre/pcreinal.h
    ! do not redefine BOOL type and TRUE / FALSE definitions

  * harbour/source/hbpcre/_hbpcreg.c
    * remap memory functions to Harbour core ones

  * harbour/source/hbpcre/pcrevutf.c
  * harbour/source/hbpcre/pcreoutf.c
  * harbour/source/hbpcre/pcrecomp.c
    * pacify warnings when utf8 mode is disabled

  * harbour/source/hbpcre/pcrefind.c
    * casting

  * harbour/contrib/gtwvg/wvtutils.c
  * harbour/contrib/rddads/adsfunc.c
    * cleanup warnings

+++ harbour/source/hbpcre/pcrefind.c
@@ -152,7 +152,7 @@ makes things a lot slower. */
 
 for (;;)
   {
-  if (top <= bot) return -1;
+  if (top <= bot) return (unsigned int) -1;
   mid = (bot + top) >> 1;
   if (c == (ucp_table[mid].f0 & f0_charmask)) break;
   if (c < (ucp_table[mid].f0 & f0_charmask)) top = mid;
